requirements.txt
google-adk
google-cloud-bigquery
google-cloud-aiplatform
python-dotenv
